+++ src/main/java/org/jabref/gui/StateManager.java
@@ -18,6 +18,7 @@ import javafx.concurrent.Task;
 import javafx.scene.Node;
 
 import org.jabref.gui.util.CustomLocalDragboard;
+import org.jabref.gui.util.DialogWindowState;
 import org.jabref.gui.util.OptionalObjectProperty;
 import org.jabref.logic.search.SearchQuery;
 import org.jabref.model.database.BibDatabaseContext;
@@ -56,6 +57,8 @@ public class StateManager {
     private final EasyBinding<Boolean> anyTaskRunning = EasyBind.reduce(backgroundTasks, tasks -> tasks.anyMatch(Task::isRunning));
 
     private final EasyBinding<Double> tasksProgress = EasyBind.reduce(backgroundTasks, tasks -> tasks.filter(Task::isRunning).mapToDouble(Task::getProgress).average().orElse(1));
+
+    private final DialogWindowState duplicateResolverDialogWindowState = new DialogWindowState();
 
     public StateManager() {
         activeGroups.bind(Bindings.valueAt(selectedGroups, activeDatabase.orElse(null)));
@@ -146,5 +149,9 @@ public class StateManager {
 
     public EasyBinding<Double> getTasksProgress() {
         return tasksProgress;
+    }
+
+    public DialogWindowState getDuplicateResolverDialogWindowState() {
+        return duplicateResolverDialogWindowState;
     }
 }